Unreviewed fix of ref-tests on GTK+. If a driver is never started, _xvfb_process is never set.
* Scripts/webkitpy/layout_tests/port/gtk.py:

Canonical link: https://commits.webkit.org/88884@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100350 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/layout_tests/port/gtk.py
@@ -57,9 +57,10 @@ class GtkDriver(WebKitDriver):
 
     def stop(self):
         WebKitDriver.stop(self)
-        # FIXME: This should use Executive.kill_process
-        os.kill(self._xvfb_process.pid, signal.SIGTERM)
-        self._xvfb_process.wait()
+        if getattr(self, '_xvfb_process', None):
+            # FIXME: This should use Executive.kill_process
+            os.kill(self._xvfb_process.pid, signal.SIGTERM)
+            self._xvfb_process.wait()
 
 
 class GtkPort(WebKitPort):
